Move ActiveRecord::ConnectionAdapters::ConnectionPool::Queue to separate file
